Api/LittleLemon-api/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djoser>=2.2.0
drf-spectacular>=0.26.0
django-filter>=23.2
